Introducing experimental consumer error handling
Provide a Spring specific general mechanism to handle consumer errors
regardless of the subscription type. Currently for exclusive and failover
subscriptions, Pulsar Java client does not allow the applicaitons to
use a DLQ. This feature allows the applications to set a PulsarConsumerErrorHandler
that takes a Backoff and PulsarMessageRecoverer to retry and recover the failed message.

This is an initial commit for this feature and more changes in this area will follow.

Resolves https://github.com/spring-projects-experimental/spring-pulsar/issues/28