Remove Brave from Java Platform (#589)
* Remove Brave from java platform

This commit removes Brave from being managed directly in the Java
Platform. Because the Brave version is already handled by Micrometer
Tracing BOM (which we import) there is no need to manage this
directly.

* Add observations back in sample failover app

This commit adds back in the observation related dependencies into the
sample-failover-custom-router sample app. They were removed temporarily
as Spring Boot 3.3.0-SNAPSHOT had a bug when they bumped to Brave 6.0.1.

Additionally:
-------------
* Add in template and listener observation-enabled properties as
Spring Boot is currently setting these to false byu default.

* Remove observation related dependencies from the sample functions app
as they were not intended to be there in the first place.