Log a warning when Lambda customizer used
This adjusts the `PulsarTemplate` to log a warning when a user provides
a `ProducerBuilderCustomizer` that is implemented as a Java Lambda.
This is important as the customizer is used as part of the producer
cache key and if not implemented properly will effectively disable
producer caching and write performance will degrade.

See #809