Update to Pulsar 3.3.2 (#875)
This commit updates the version of Pulsar to 3.3.2.

Also, in Pulsar 3.3.2 the schema validation of an outgoing message value
happens later than it did previously. This requires the PulsarTemplate
to widen the try/catch net so that when this happens the producer is
closed properly. The backing change in Pulsar 3.3.2 can be seen here
https://github.com/apache/pulsar/commit/f3c177e2243e26a7849feb91dbed9fec4c5723c0#diff-095bc2359e03726e031d8c2f210560c6a3218f48a969b0f75a2482e25cd54744R69