Update to Spring Javaformat / checkstyle (prepare for JSpecify)
This commit updates to Javaformat to 0.0.47 and the Checkstyle
tools to 10.25.0.

The `SpringAnnotationLocationCheck` has also been added to the
checkstyle rules in order to prepare for the migration to Jspecify
for nullability constraints.

Also, various code has been updated to abide by the new format and
checkstyle rules in the updated versions.

See #1150

Signed-off-by: Chris Bono <chris.bono@gmail.com>