Reinstate 'pubishArtifacts' Gradle task
This commit adds back in the Gradle `PublishArtifactsPlugin` and updates
the CI to use it. The reason is that the spring-pulsar-docs module never
goes to Maven Central but it does need to always go to Artifactory.

Signed-off-by: Chris Bono <chris.bono@gmail.com>
(cherry picked from commit 5647e9695e2fb647f9b4dd1d2859d36232330f7c)