Remove Java 8 requirement from sample function app
This removes the previous requirement for the sample-signup-app to target
Java 8. It was required due to a now non-existent limitation in the Pulsar
functions client.

It also removes the previous requirement for the sample-signup-app to use
Pulsar client version `4.0.5`. It was required due to a limmitation in the
Pulsar version `4.0.6` which has since been addressed in Pulsar `4.0.6+`.
Therefore, the Pulsar version used in now derived from the Gradle platform.

Signed-off-by: onobc <chris.bono@gmail.com>